Changes to schema to loosen dependencies between Forced ID table and the Resource and Resource History tables.
