+++ hapi-fhir-jpaserver-model/src/main/java/ca/uhn/fhir/jpa/model/entity/ResourceHistoryTable.java
@@ -21,6 +21,8 @@ package ca.uhn.fhir.jpa.model.entity;
  */
 
 import ca.uhn.fhir.jpa.model.cross.ResourcePersistentId;
+import ca.uhn.fhir.model.primitive.IdDt;
+import ca.uhn.fhir.rest.api.Constants;
 import org.hibernate.annotations.OptimisticLock;
 
 import javax.persistence.*;
@@ -54,7 +56,11 @@ public class ResourceHistoryTable extends BaseHasResource implements Serializabl
 	@Column(name = "PID")
 	private Long myId;
 
-	@Column(name = "RES_ID")
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "RES_ID", nullable = false, updatable = false, foreignKey = @ForeignKey(name = "FK_RESOURCE_HISTORY_RESOURCE"))
+	private ResourceTable myResourceTable;
+
+	@Column(name = "RES_ID", nullable = false, updatable = false, insertable = false)
 	private Long myResourceId;
 
 	@Column(name = "RES_TYPE", length = ResourceTable.RESTYPE_LEN, nullable = false)
@@ -165,4 +171,35 @@ public class ResourceHistoryTable extends BaseHasResource implements Serializabl
 	public ResourcePersistentId getPersistentId() {
 		return new ResourcePersistentId(myResourceId);
 	}
+
+	public ResourceTable getResourceTable() {
+		return myResourceTable;
+	}
+
+	public void setResourceTable(ResourceTable theResourceTable) {
+		myResourceTable = theResourceTable;
+	}
+
+	@Override
+	public IdDt getIdDt() {
+		if (getResourceTable().getForcedId() == null) {
+			Long id = getResourceId();
+			return new IdDt(getResourceType() + '/' + id + '/' + Constants.PARAM_HISTORY + '/' + getVersion());
+		} else {
+			// Avoid a join query if possible
+			String forcedId = getTransientForcedId() != null ? getTransientForcedId() : getResourceTable().getForcedId().getForcedId();
+			return new IdDt(getResourceType() + '/' + forcedId + '/' + Constants.PARAM_HISTORY + '/' + getVersion());
+		}
+	}
+
+	@Override
+	public ForcedId getForcedId() {
+		return getResourceTable().getForcedId();
+	}
+
+	@Override
+	public void setForcedId(ForcedId theForcedId) {
+		getResourceTable().setForcedId(theForcedId);
+	}
+
 }
